Improved context menu in Conflict window
Using ContextPopUp for Conflict window as well.
Usage is easier and conforming to the general behaviour of
context menus, because there's no delay after right-click.

Moved "Open file" to be the first item in the context menu.
It saves having to move the mouse to selecct that item and
opening the file is therefore an easy right-double-click.
"Open folder" is already quickly done with a simple left-click.